Use nwb for build, lodash babel, and es6 imports
