+++ __tests__/browserifySpec.js
@@ -1,8 +1,9 @@
 'use strict';
 
-var should = require('should');
-var expect = require('expect');
-var assert = require('assert');
+import should from 'should';
+import expect from 'expect';
+import assert from 'assert';
+import ItemsJS from '../src/index';
 
 describe('itemjs general tests', function() {
 
@@ -11,7 +12,6 @@ describe('itemjs general tests', function() {
     tags: ['a', 'b', 'c', 'd'],
     actors: ['a', 'b']
   }, {
-    id: 10,
     name: 'movie2',
     tags: ['a', 'e', 'f'],
     actors: ['a', 'b']
@@ -21,43 +21,11 @@ describe('itemjs general tests', function() {
     actors: ['e']
   }]
 
-  var itemsjs = require('./../src/index')(items, {
-    aggregations: {
-      tags: {
-        size: 10
-      }
-    }
-  });
-
-  it('checks aggregations', function test(done) {
-    var result = itemsjs.aggregation({
-      name: 'tags'
-    });
-
-    assert.equal(result.data.buckets.length, 6);
-    assert.equal(result.data.buckets[0].key, 'a');
-    assert.equal(result.data.buckets[0].doc_count, 3);
-    done();
-  });
-
-  it('checks aggregation with undefined aggregation', function test(done) {
-
-    try {
-      itemsjs.aggregation({
-        name: 'colors'
-      });
-    } catch (e) {
-      assert.ok(e.message.indexOf('Please define aggregation') !== -1);
-      done();
-    }
-  });
+  var itemsjs = ItemsJS(items);
 
   it('makes search', function test(done) {
     var result = itemsjs.search();
     assert.equal(result.data.items.length, 3);
-    assert.equal(result.data.items[0].id, 1);
-    assert.equal(result.data.items[1].id, 10);
-    assert.equal(result.data.items[2].id, 2);
     done();
   });
 
@@ -81,19 +49,9 @@ describe('itemjs general tests', function() {
     done();
   });
 
-  it('makes search with prefilter', function test(done) {
-    var result = itemsjs.search({
-      prefilter: function(items) {
-        return items.slice(0, 1);
-      }
-    });
-    assert.equal(result.data.items.length, 1);
-    done();
-  });
-
   it('makes search with aggregation filters', function test(done) {
 
-    var itemsjs = require('./../src/index')(items, {
+    var itemsjs = ItemsJS(items, {
       aggregations: {
         tags: {},
         actors: {}
@@ -129,7 +87,7 @@ describe('itemjs general tests', function() {
   });
 
   it('makes aggregations when configuration supplied', function test(done) {
-    var itemsjs = require('./../src/index')(items, {
+    var itemsjs = ItemsJS(items, {
       aggregations: {
         tags: {
           type: 'terms',
@@ -139,103 +97,6 @@ describe('itemjs general tests', function() {
       }
     });
     var result = itemsjs.search({});
-    assert.equal(result.data.items.length, 3);
-    assert.equal(result.data.aggregations.tags.name, 'tags');
-    assert.equal(result.data.aggregations.tags.buckets.length, 6);
-    done();
-  });
-
-  it('makes search aggregations and keep original configuration', function test(done) {
-
-    var items = [{
-      name: 'movie1',
-      tags: ['a', 'b', 'c', 'd'],
-      actors: ['a', 'b']
-    }, {
-      id: 10,
-      name: 'movie2',
-      tags: ['a', 'e', 'f'],
-      actors: ['a', 'b', 'c', 'd', 'e', 'f']
-    }, {
-      name: 'movie3',
-      tags: ['a', 'c'],
-      actors: ['e']
-    }]
-
-    var itemsjs = require('./../src/index')(items, {
-      aggregations: {
-        tags: {
-          type: 'terms',
-          size: 10,
-        },
-        actors: {
-          type: 'terms',
-          size: 10,
-        }
-      }
-    });
-    var result = itemsjs.search({
-      filters: {
-        tags: ['c']
-      }
-    });
-    assert.equal(result.data.items.length, 2);
-    assert.equal(result.data.aggregations.tags.name, 'tags');
-    assert.equal(result.data.aggregations.tags.buckets.length, 4);
-
-    var result = itemsjs.aggregation({
-      name: 'actors',
-      per_page: 10
-    });
-
-    assert.equal(result.data.buckets.length, 6);
-    done();
-  });
-
-  it('makes aggregations when configuration supplied and input provided', function test(done) {
-    var itemsjs = require('./../src/index')(items, {
-      aggregations: {
-        tags: {
-          type: 'terms',
-          size: 10,
-          title: 'Tags'
-        }
-      }
-    });
-    var result = itemsjs.search({
-      filters: {
-        tags: ['a']
-      }
-    });
-    assert.equal(result.data.items.length, 3);
-    assert.equal(result.data.aggregations.tags.name, 'tags');
-    assert.equal(result.data.aggregations.tags.buckets.length, 6);
-
-    var result = itemsjs.search({
-      filters: {
-        tags: ['a', 'z']
-      }
-    });
-    assert.equal(result.data.items.length, 0);
-    assert.equal(result.data.aggregations.tags.name, 'tags');
-    assert.equal(result.data.aggregations.tags.buckets.length, 0);
-
-    var itemsjs = require('./../src/index')(items, {
-      aggregations: {
-        tags: {
-          type: 'terms',
-          size: 10,
-          conjunction: false,
-          title: 'Tags'
-        },
-      }
-    });
-
-    var result = itemsjs.search({
-      filters: {
-        tags: ['a', 'c']
-      }
-    });
     assert.equal(result.data.items.length, 3);
     assert.equal(result.data.aggregations.tags.name, 'tags');
     assert.equal(result.data.aggregations.tags.buckets.length, 6);
@@ -254,7 +115,7 @@ describe('itemjs general tests', function() {
       tags: 'a',
     }];
 
-    var itemsjs = require('./../src/index')(items, {
+    var itemsjs = ItemsJS(items, {
       aggregations: {
         tags: {
           type: 'terms',
@@ -280,7 +141,7 @@ describe('itemjs general tests', function() {
       name: 'movie3',
     }];
 
-    var itemsjs = require('./../src/index')(items, {
+    var itemsjs = ItemsJS(items, {
       aggregations: {
         tags: {
           type: 'terms',
@@ -308,7 +169,7 @@ describe('itemjs general tests', function() {
       name: 'movie3',
     }];
 
-    var itemsjs = require('./../src/index')(items, {
+    var itemsjs = ItemsJS(items, {
       searchableFields: ['name', 'tags']
     });
     var result = itemsjs.search({
