+++ nwb.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  type: 'web-module',
+  babel: {
+    plugins: ['lodash'],
+  },
+  npm: {
+    esModules: true,
+    umd: {
+      global: 'itemsjs',
+      entry: './src/index.js',
+      externals: {}
+    }
+  }
+}
