+++ package.json
@@ -2,13 +2,13 @@
   "name": "itemsjs",
   "version": "1.0.39",
   "description": "Created to perform fast search on small json dataset (up to 1000 elements).",
-  "main": "index.js",
+  "main": "lib/index.js",
+  "module": "es/index.js",
   "scripts": {
-    "test": "mocha tests/*",
-    "browserify": "browserify index.js -v -s itemsjs -o dist/itemsjs.js",
-    "minify": "minify dist/itemsjs.js -o dist/itemsjs.min.js",
-    "dist": "npm run browserify && npm run minify",
-    "lodash": "lodash include=some,forEach,map,mapKeys,mapValues,every,includes,intersection,filter,keys,flatten,transform,orderBy -o lib/lodash.js -p"
+    "build": "nwb build-web-module && rm -rf dist && mv umd dist",
+    "clean": "nwb clean-module",
+    "test": "jest",
+    "prepublish": "npm run build"
   },
   "author": "Matt Rzepa",
   "license": "Apache-2.0",
@@ -18,10 +18,9 @@
     "lunr": "^1.0.0"
   },
   "devDependencies": {
-    "babel-minify": "^0.2.0",
-    "browserify": "^14.3.0",
     "expect": "^1.20.2",
-    "lodash-cli": "^4.17.4",
+    "jest": "^23.5.0",
+    "nwb": "^0.23.0",
     "should": "^11.2.1",
     "sinon": "^2.1.0"
   },
@@ -46,5 +45,10 @@
   "bugs": {
     "url": "https://github.com/itemsapi/itemsjs/issues"
   },
-  "homepage": "https://github.com/itemsapi/itemsjs#readme"
+  "homepage": "https://github.com/itemsapi/itemsjs#readme",
+  "babel": {
+    "presets": [
+      "env"
+    ]
+  }
 }
